llm_api/pyproject.toml
[project]
name = "llm-api"
version = "0.1.0"
description = "LLM API for study support"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115.0,<1.0.0",
  "uvicorn[standard]>=0.30.0,<1.0.0",
  "pydantic>=2.7.0,<3.0.0",
  "openai>=1.40.0,<2.0.0",
  "python-dotenv>=1.0.1,<2.0.0",
  "httpx>=0.27.0,<1.0.0",
]

[tool.uv]
package = false

[build-system]
requires = ["hatchling>=1.25.0"]
build-backend = "hatchling.build"
